==> mvp/notebooks/3-Dataframe setup.ipynb <==
# Clinical trials: Dataframe setup


import time
import json
import os
import pandas as pd

# Path for json folder & file
path_to_json_file = os.path.abspath('../data/json/')

# json file
json_file = '/all_trials_json' #name json file  

## Import json file in a dataframe/sframe

# Import json into a dataframe
file = '{}{}.json'.format(path_to_json_file, json_file)

# breaks with large json file
df = pd.read_json(file)

## Basic data cleaning

#Check data
df.info()

df.head()

# remove \n values
df = df.replace(r'\n',' ', regex=True) 

# df.set_index('nct_id', inplace=True)

df.columns

# Rename columns
df.columns = ['id', 
              'submission_date',
              'source', 
              'brief_title', 
              'condition', 
              'mesh_term_condition', 
              'mesh_term_intervention', 
              'full_description', 
              'summary', 
              'city', 
              'country', 
              'zip']

df.head()

## Add new date columns

# Create new column: study_first_submitted as dates
%time df['full_date'] = pd.to_datetime(df['submission_date'])

# Create new column: dates as year
df['year'] = df['full_date'].dt.year

# Delete submission date column
df.drop('submission_date', axis=1, inplace=True)

# Sort records by date
df = df.sort_values(by ='full_date')

df.head()

df.info()

## Subset of data

def filter_data(year, df):
    # Select data since 2008
    df = [df['year'] > year]

# Select data since 2008
# df = df[df['year'] > 2007]
# df.info()

filter_data(2008, df)

df.mesh_term_condition.nunique()

df.condition.nunique()

df.mesh_term_intervention.nunique()

## Export dataframe as csv

def export_data_csv(df, path_to_csv = os.path.abspath('../data/csv/'), csv_file = '/clean_data.csv'):
    try:
        os.mkdir(path_to_csv)
        print('{} created'.format(csv))
    except IOError as e:
        print(e)
        pass
    
    df.to_csv(path_to_csv + csv_file)

# Export dataframe as csv file

# path_to_csv = os.path.abspath('../data/csv/')

# try:
#     os.mkdir(path_to_csv)
#     print('{} created'.format(csv))
# except IOError as e:
#     print(e)
#     pass

export_data_csv(df)

csv_file = '/clean_data.csv'
df.to_csv(path_to_csv + csv_file)
