==> mvp/notebooks/Script 3 - df setup, export working dataset.ipynb <==
# Prepare working db
- Load data
- df cleaning
- Create label for recruting status
- Export full dataset to sqlite
- Export dataset to postgres

import json, os
import numpy as np
import pandas as pd
import sqlite3
import shutil
import matplotlib.pyplot as plt
from sqlalchemy import create_engine
# import nltk

## Load json file with parsed information

# paths
path_to_json_file = os.path.abspath('../data/json/')

### json file
# json_file = '/all_parsed_data_json' #name json file  
json_file = '/sample_json' #sample json file for testing  

file = '{}{}.json'.format(path_to_json_file, json_file)

# load json file
df = pd.read_json(file)

# Display full dataframe
# pd.set_option('display.max_rows', 50)
# pd.set_option('display.max_columns', 50)
# pd.set_option('display.width', 1000)

# Check data
df.shape

# Remove all_files folder
path_to_all_files = os.path.abspath('../data/all_trials//')

def remove_extra_folders(folder):
    try:
        shutil.rmtree(folder)
        print("\nunzip folder deleted")
    except IOError as e:
        print(e)
        pass

%time remove_extra_folders(path_to_all_files)

## Format dataset

df.head()

# function to change column type

def data_types(dataframe, cols = [], to_type = ''):
    for col in cols:
        dataframe[col] = df[col].astype(to_type)
        

df['verification_date'].sample(10)

## Change date types, extract years

columns_dates = ['study_first_submitted', 'last_update_submitted', 'verification_date']
data_types(df, columns_dates, 'datetime64')

# Extract year 
df['year_submitted'] = df['study_first_submitted'].dt.year
df['year_last_updated'] = df['last_update_submitted'].dt.year
df['verification_year'] = df['verification_date'].dt.year

# Remove unnecesary columns
# df.drop(columns_dates, axis = 1, inplace=True)

# columns_to_drop = ['study_first_posted', 'last_update_posted', 'verification_date']
# df.drop(columns_to_drop, axis = 1, inplace=True)

# Remove /n in all df
df = df.replace(r'\n',' ', regex=True)

# Create new column for all text
df['all_text'] = df['source'] + ' ' + df['brief_title'] + ' ' + df['condition'] + ' ' + df['condition_browse/mesh_term'] + ' '+ df['intervention_browse/mesh_term'] + ' '+ df['detailed_description/textblock'] + ' ' + df['brief_summary/textblock']

# All_text in lowercase
df['all_text'] = df['all_text'].str.lower()

# remove extra whitespace
df.all_text = df.all_text.replace('\s+', ' ', regex=True)


# Remove extra white space in summary
df['brief_summary/textblock'] = df['brief_summary/textblock'].replace('\s+', ' ', regex=True)
df['brief_summary/textblock'][0]

# # add url columns
url_string = 'https://clinicaltrials.gov/ct2/show/'
df['url'] = url_string + df['nct_id'].astype(str)

## Recruiting status
- Create criteria for filtering recruiting status
- Create new column with labels

''' Recruiting status 
0 - Not recruting, including all records not updated 
1 - Possibily recruiting. Define time range for category 
2 - Recruiting

'''

df.overall_status.unique()

# Selecting open trials
# https://clinicaltrials.gov/ct2/help/glossary/recruitment-status

recruiting = df[(df['overall_status'] == "Recruiting") | \
                (df['overall_status'] == "Not yet recruiting") | \
                (df['overall_status'] == "Available for expanded access")]

# Check recruiting studies
print('Total of trials classified as recruiting: {}'.format(len(recruiting)))

print('Trials not verified in the last two years 2017-2019: {}'.\
      format(len(recruiting[recruiting['verification_year'] < 2017])))

recruiting.overall_status.unique()

# Trials by verification date

'''Unknown: A study with a status of Recruiting, Not yet recruiting, 
or Active, not recruiting and whose status has not been verified within the past 2 years. 

** Studies with an Unknown recruitment status are considered open studies or closed studies, 
depending on their last known recruitment status**

'''
recruiting.boxplot(column='verification_year', by='overall_status', \
                   figsize=(14,9), showfliers=False, meanline=True,\
                   whiskerprops = dict(linestyle='-.', linewidth=2))

# recruiting.boxplot(column='year_last_updated')
# recruiting['verification_year'].plot(kind='box', figsize=(10,7))
# recruiting['year_last_updated'].plot(kind='box', figsize=(10,7))

# Create new column with recruiting status based on verification date

# Function to assign recruiting labels
options = [
    (df['verification_year'] < 2017),
    (df['verification_year'] > 2017) &
    (df['year_last_updated'] > 2017) & 
    (df['overall_status'] == "Recruiting") | \
                (df['overall_status'] == "Not yet recruiting") | \
                (df['overall_status'] == "Available for expanded access")]


labels = [0, 1]
df['recruiting_labels'] = np.select(options, labels, default=0)
df['recruiting_labels'].value_counts()


all_recruiting = df[df['recruiting_labels'] == 1].sort_values('study_first_submitted')
print('Verification years: {}'.format(all_recruiting['verification_year'].unique()))
print('Verification dates: {}'.format(all_recruiting['verification_date'].unique()))
print('Year last udpated: {}'.format(all_recruiting['year_last_updated'].unique()))

all_recruiting['verification_date'].unique()

# all_recruiting.boxplot(column='recruiting_labels', by='sverification_date', \
#                    figsize=(14,9), showfliers=False, meanline=True,\
#                    whiskerprops = dict(linestyle='-.', linewidth=2))


# Connect to sqliteDB

# Create folder

path_to_working_datasets = os.path.abspath('../data/working_data')

try: 
    os.mkdir(path_to_working_datasets)
except:
    pass
    print('Error')

# Create and connect to db
conn = sqlite3.connect('../data/working_data/working-database.db')
conn


# upload df to database: define schema
%time df.to_sql('all_trials', con=conn)

# get list of indexes
pd.read_sql_query("PRAGMA index_list(all_trials);", conn)

# test 1. Filter by condition
%time search_query = pd.read_sql_query("SELECT * from all_trials WHERE all_text LIKE '%breast cancer%';", conn)
search_query.head()

# test 2. Filter by recruiting status = 1
%time search_query = pd.read_sql_query("SELECT * from all_trials WHERE recruiting_labels == 1;", conn)
len(search_query)

# Create index on all_text column
c = conn.cursor()
%time c.execute("CREATE INDEX idx1 ON all_trials(all_text)")
conn.commit()

# get list of indexes
pd.read_sql_query("PRAGMA index_list(all_trials);", conn)

# test new query
%time search_query = pd.read_sql_query("SELECT * from all_trials WHERE all_text LIKE '%breast cancer%';", conn)
search_query.head()

%time new_query = pd.read_sql_query("SELECT nct_id from all_trials WHERE all_text LIKE '%breast cancer%';", conn).head()
new_query.head()

'''
Optimize performance
https://medium.com/@JasonWyatt/squeezing-performance-from-sqlite-indexes-indexes-c4e175f3c346
'''

# Export df to postgres

%time postgres_df = pd.read_sql_query("SELECT * from all_trials;", conn)

engine = create_engine('postgresql://cms@localhost:5432/ClinicalTrials')

engine

postgres_df.to_sql('trials_12_1_19', engine)

_____

## Test queries speed in sqlite

# Check data types
# pd.read_sql_query("PRAGMA table_info(all_trials);", conn)

# Time query with WHERE condition
%time search_query = pd.read_sql_query("SELECT * from all_trials WHERE all_text LIKE '%breast cancer%';", conn)
search_query.head()

# Time query with WHERE condition, return nct_id
%time filtered_query = pd.read_sql_query("SELECT nct_id from all_trials WHERE all_text LIKE '%breast cancer%';", conn)
filtered_query.head()

# drop original index
%time c.execute("DROP INDEX ix_all_trials_index")
conn.commit()


## Test queries after creating index

# Time query with WHERE condition after creating index
%time optimized_query = pd.read_sql_query("SELECT * from all_trials WHERE all_text LIKE '%breast cancer%';", conn)
optimized_query.head()

# Return only nct_id
%time new_query = pd.read_sql_query("SELECT nct_id from all_trials WHERE all_text LIKE '%breast cancer%';", conn).head()
new_query.head()


## Preprocess dataset
- Tokenize, Lemmatize / Stem
- Remove stopwords

from nltk.stem import PorterStemmer
from nltk import WordNetLemmatizer

# porter_stemmer = PorterStemmer()
lemmatizer = WordNetLemmatizer() 

# Stemming
# def stem_sentences(text):
#     tokens = text.split()
#     stemmed_tokens = [porter_stemmer.stem(token) for token in tokens]
#     return ' '.join(stemmed_tokens)

# df['stems'] = df['all_text'].apply(stem_sentences)

# Lemmatizing
def lemm_sentences(text):
    tokens = text.split()
    lemm_tokens = [lemmatizer.lemmatize(token) for token in tokens]
    return ' '.join(lemm_tokens)

df['lemmas'] = df['all_text'].apply(lemm_sentences)

# Remove stop words [not needed with countvectorizer]
# stop_words = stopwords.words('english')
# df['tokens'] = df['lemmas'].apply(lambda x: [item for item in x.split() if item not in stop_words])

df.lemmas[0:5]

## Bags of words

# Count Vectorizer
from sklearn.feature_extraction.text import CountVectorizer
import re

pat_numbers = lambda x: re.sub(r'(\d)+', '', x.lower())

cv = CountVectorizer(stop_words='english',
                     preprocessor = pat_numbers,
                     max_features = 1800,
                     lowercase = True,
                     max_df = 0.8,
                     ngram_range = (1, 3))

docs = list(df['lemmas'])

X = cv.fit_transform(docs)

len(cv.vocabulary_)

word_counts = pd.DataFrame(X.toarray(), columns=cv.get_feature_names())

# add nct_id to word_counts df
nct_id = df['nct_id']

df_word_counts = pd.merge(nct_id, word_counts, left_index=True, right_index=True)

df_word_counts.head()

_______

# Check all dataframes in space
%who DataFrame

# upload count_vectorizer to database
# sqlite limit = 2000
df_word_counts.to_sql('word_counts', con=conn)

df_counts = pd.read_sql_query("SELECT * from word_counts;", conn)
df_counts.head()

# List databases
pd.read_sql_query("SELECT name FROM sqlite_master WHERE type='table';", conn)

conn.close()

--------

## TF-IDF

from sklearn.feature_extraction.text import TfidfTransformer

transformer = TfidfTransformer()
tweights = transformer.fit_transform(X)
tweights

# turn weights data into a dataframe
tf = pd.DataFrame(tweights.toarray(), columns=cv.get_feature_names())

# Top terms by average tf-idf weight
weights = np.asarray(tweights.mean(axis=0)).ravel().tolist()
weights_df = pd.DataFrame({'term': cv.get_feature_names(), 'weight': weights})
weights_df.sort_values(by='weight', ascending=False).head(10)

# Create new dataframe with nct-id and merge wth tf by index
pd.set_option('display.max_columns', None) 

df_tf = df['nct_id']
df_tf.head()

df_tf = pd.merge(df_tf, tf, left_index=True, right_index=True)

df_tf.head()

## Test merged dataframes with tf-idf results

# import random

# def compare_results():
#     row = random.randint(0,3000)
#     df_tf.drop('nct_id', axis = 1)
#     print('Random row: {}'.format(row))
#     return df_tf.loc[row], tf.loc[row]

# compare_results()

## Calculate correlation between docs

# calculate similary
similarity = tweights * tweights.T

# Save all data in a dataframe
df_docs_similarity = pd.DataFrame(similarity.toarray())

<b>To-do: Give a NTC-ID record, find similar documents
& return dataframe with results</b>

# Find more similar documents of a given record

def find_similar_docs(record, rate):
    trials_id = []
    treshold = rate
    similar_index = df_docs_similarity.iloc[record][df_docs_similarity.iloc[record] > treshold].index
    
    for i in similar_index.values:
        trials_id.append(i)
        
    print('{} similar trials with treshold {}'.format(len(trials_id), treshold))
    return df.iloc[trials_id]


find_similar_docs(0, 0.50)
