==> mvp/notebooks/Script 3b - df export to elasticsearch.ipynb <==
## Exporting clinical trials db to elasticsearch

import pandas as pd
import sqlite3
from elasticsearch import Elasticsearch
import request, json, os

## Connect to sqlite

# connect to sqlite database
conn = sqlite3.connect('../data/working_data/working-database.db')
conn

# get list of tables
pd.read_sql_query("select name from sqlite_master where type = 'table';", conn)


df = pd.read_sql_query("select * from all_trials;", conn)

df.head()

## Create json file

dicts = df.to_dict(orient='records')

dicts[0]

# export json file to elasticsearch folder

### Connect to elasticsearch server
- https://github.com/Moshe/elasticsearch_loader

import elasticsearch_loader

# Create index
# curl -XPUT "http://localhost:9200/IndexName/TypeName"

# elasticsearch_loader --index clinicaltrials --type incident json ../data/elasticsearch/*.json
