==> mvp/notebooks/Script 5 - Creating Pubmed db.ipynb <==
## Import data into pySpark
- https://github.com/titipata/pubmed_parser/wiki/Download-and-preprocess-MEDLINE-dataset

- MEDLINE BULK DOWNLOAD -
wget ftp://ftp.ncbi.nlm.nih.gov/pubmed/baseline/*.gz

- MEDLINE UPDATES -
wget ftp://ftp.ncbi.nlm.nih.gov/pubmed/updatefiles/*.gz
- Change JAVA version: https://kodejava.org/how-do-i-set-the-default-java-jdk-version-on-mac-os-x/

import os
import pandas as pd
from glob import glob
import pubmed_parser as pp
import findspark
from pyspark.sql import SparkSession
from pyspark import SparkConf, SparkContext
from pyspark.sql import Row
from pyspark.sql import SQLContext
findspark.init()

# Create spark session and variables
sc = SparkContext("local", "medline db")
conf = SparkConf()
spark = SparkSession.builder.\
    config(conf=conf)\
#     .getOrCreate()

sqlContext = SQLContext(sc)

## Import medline files into a spark dataframe
- https://github.com/titipata/pubmed_parser/wiki/Download-and-preprocess-MEDLINE-dataset

medline_files_rdd = sc.parallelize(glob('../data/pubmed/*.gz'))

medline_files_rdd

parse_results_rdd = medline_files_rdd.\
    flatMap(lambda x: [Row(file_name=os.path.basename(x), **publication_dict) 
                       for publication_dict in pp.parse_medline_xml(x)])

parse_results_rdd

medline_df = parse_results_rdd.toDF()

# Process upates and deletes
from pyspark.sql import Window
from pyspark.sql.functions import rank, max, sum, desc

window = Window.partitionBy(['pmid']).orderBy(desc('file_name'))

windowed_df = medline_df.select(
    max('delete').over(window).alias('is_deleted'),
    rank().over(window).alias('pos'),
    '*')

medline_lastview = windowed_df.where('is_deleted = False and pos = 1')

type(medline_lastview)

## Explore db
- https://medium.com/@aieeshashafique/exploratory-data-analysis-using-pyspark-dataframe-in-python-bd55c02a2852

medline_lastview.printSchema()

# Select columns
columns = ['abstract', 'file_name', 'journal', 'keywords', 'mesh_terms', 'pubdate'] 
df = medline_lastview[columns]


df.printSchema()

# df.collect()

# df = df.toPandas()

## Extracting medline citations

import medic

wget "$URL?db=PubMed&retmax=99&term=$QUERY" -O - 2> /dev/null \
| grep "^<Id>" \
| sed -E 's|</?Id>||g' \
| cut -f3 \
> pmids.txt

## Calculate TF-IDF
- https://github.com/titipata/pubmed_parser/wiki/Download-and-preprocess-MEDLINE-dataset

## Classify with Fastest
- https://www.futurice.com/blog/classifying-text-with-fasttext-in-pyspark
- https://fasttext.cc/docs/en/supervised-tutorial.html

# Topic modeling
- https://nbviewer.jupyter.org/github/chambliss/Notebooks/blob/master/Word2Vec_News_Analysis.ipynb
