==> mvp/notebooks/Script 6 - Connecting and querying aact postgres db.ipynb <==
## Create working aact db

import pandas as pd
from sqlalchemy import create_engine
import psycopg2
%load_ext sql

## DB connection

# connecting to aact database
%sql postgresql://cms:postgres@localhost:5432/aact

# Save query in a dataframe
%time df = %sql select * from trials
%time df = df.DataFrame()

### Alternative connection

# Local postgres

host = 'postgresql://cms@localhost:5432/' 

db = 'aact'
connection = host_cpj + db

engine = create_engine(connection)

%time sq = pd.read_sql_query("SELECT * from trials", engine)

# sqlite df to postgres table
# sq.to_sql('trials_12_1_19', engine)

# SELECT table_name FROM information_schema.tables \
# WHERE table_schema = 'ctgov'\
# order by table_name asc;

# select * from ctgov.all_keywords
# where all_keywords.names like '%celiac%';

# Save query in a dataframe
# %time df = %sql select * from ctgov.studies
# %time df = df.DataFrame()

# SELECT
# (select count(*) as count_mesh_terms from ctgov.mesh_terms), 
# (select count(*) as count_conditions from ctgov.conditions);

## Create new trials db
- Info on AACT schema: https://aact.ctti-clinicaltrials.org/data_dictionary
- Main table: ctgov.studies, ctgov.brief_summaries, ctgov.brief_summaries, ctgov.interventions, ctgov.conditions
- all_conditions: mesh_terms & conditions

# Check number of records in different tables

%time counts = %sql SELECT \
(select count(*) as all_trials from trials), \
(select count(*) as conditions from ctgov.conditions),\
(select count(*) as browse_conditions from ctgov.browse_conditions),\
(select count(*) as all_conditions from ctgov.all_conditions),\
(select count(*) as browse_interventions from ctgov.browse_interventions),\
(select count(*) as brief_summaries from ctgov.brief_summaries)

counts

### Create trials table

# %sql 
# CREATE TABLE trials as
# SELECT 
#     nct_id, 
#     phase, 
#     study_first_submitted_date, 
#     study_first_submitted_qc_date,
#     "study_first_submitted_qc_date"::date - "study_first_submitted_date"::date AS submitted_to_qc,
#     study_first_posted_date,
#     results_first_submitted_date is not null as results,
#     study_type,
#     overall_status,
#     why_stopped is not null as stopped,
#     why_stopped,
#     has_expanded_access is true as has_expanded_access,
#     is_fda_regulated_drug,
#     is_fda_regulated_device,
#     is_unapproved_device,
#     official_title,
#     acronym,
#     source
#     FROM ctgov.studies;

### Add column with brief_description

# Find trials with null values
%sql SELECT \
   nct_id, \
   'not in summaries' as note \
FROM \
   trials \
EXCEPT \
   SELECT \
    nct_id, \
    'not in summaries' as note \
  FROM \
    ctgov.brief_summaries

# Create column for description
ALTER TABLE trials
ADD COLUMN description text;

# Select brief_summaries to trials table
UPDATE trials AS t1 
SET description = t2.brief_summaries
FROM brief_summaries AS t2
WHERE t1.nct_id = t2.nct_id

# JOIN
# %sql select * from trials as t1 \
# left join ctgov.brief_summaries as t2 \
# on t1.nct_id = t2.nct_id

# Check
select count(*) from trials
where description is null

## Add column with mesh_terms

# Create column for mesh?terms
ALTER TABLE trials
ADD COLUMN mesh_terms text;

# Join all mesh_terms in new table
CREATE TABLE all_mesh_terms as
select 
	nct_id,
	STRING_AGG (downcase_mesh_term, ',') mesh_terms
from 
	ctgov.browse_conditions
group by 
	nct_id;


#update trials table with all_mesh_terms
UPDATE trials AS t1 
SET mesh_terms = t2.mesh_terms
FROM all_mesh_terms AS t2
WHERE t1.nct_id = t2.nct_id

### Create column with all interventions

# Create column for interventions
ALTER TABLE trials
ADD COLUMN interventions text;

# Join all interventions in new table
CREATE TABLE all_interventions as
select 
	nct_id,
	STRING_AGG (downcase_mesh_term, ',') mesh_terms
from 
	ctgov.browse_interventions
group by 
	nct_id;

#update trials table with all_interventions
UPDATE trials AS t1 
SET interventions = t2.mesh_terms
FROM all_interventions AS t2
WHERE t1.nct_id = t2.nct_id

## Add column with recruiting status

# Recruiting_status: Recruiting, Not yet recruiting, Available for expanded access. 
# Create column 
ALTER TABLE trials
ADD COLUMN recruiting_status text;

# Update recruiting_status based on condition
UPDATE trials
SET recruiting_status = CASE
WHEN overall_status = 'Recruiting' or overall_status = 'Not yet recruiting' THEN 1 ELSE 0 END

# Change data type
ALTER TABLE trials
ALTER COLUMN recruiting_status TYPE bool USING recruiting_status::boolean;

## List of search terms
- Get distinct conditions and interventions in dataset

# Get list of all interventions & conditions
select downcase_mesh_term, count(downcase_mesh_term) from ctgov.browse_interventions
group by downcase_mesh_term order by count desc;

select downcase_mesh_term, count(downcase_mesh_term) from ctgov.browse_conditions
group by downcase_mesh_term order by count desc;

# Evaluate results by queries
SELECT 
(select count(*) as conditions_hiv_counts from ctgov.conditions where downcase_name = 'hiv'),

(select count(*) as browse_conditions_hiv_counts from ctgov.browse_conditions where downcase_mesh_term like '%hiv%'),

(select count(*) as conditions_celiac_counts from ctgov.conditions where downcase_name = 'celiac'),

(select count(*) as browse_conditions_celiac_counts from ctgov.browse_conditions where downcase_mesh_term like '%celiac%')
